feat: replace socket.io with native websockets
Removes the overhead socket.io introduces. Modern clients already support WebSockets, so a fallback is not needed. If a fallback is somehow still needed for certain edge cases, a custom socket server/client implementation (or even socket.io) can easily be provided to replace the default implementation.

BREAKING CHANGE: API for initializing Server and Client was changed. It does not need a socket.io instance anymore.