fix(client): first subscription sometimes not receiving value
Workaround. When subscribing to a path via "once/on" where there was already data set locally before
(via "set"), the listener would never be triggered, because the received data from server is not
different to the locally stored data. This workaround will now only store data when subscribed to
the path and delete data when unsubscribing.

fix #19